Welp; that wasn't the issue.
So turning that back and leaving everything else as I found it.

Trying to build on my local and with both Guile 2.2 and 3.0 both
result in the same thing, whether 0.4 or 0.4.1.

I must've built this before, right? If that's the first version I
committed?

I dunno. I sent a message to the mailing list so we'll have to see.